refactor(chat): extract chat_manager/deps_config.rs (#292)
Task 11 of the split plan: CmDeps and its ConfigManagerDeps impl
move into their own file, owning their own construction via a
build() helper the constructor now calls.
